% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claytonCopulaDF.R
\name{ClaytonLogLik}
\alias{ClaytonLogLik}
\title{ClaytonLogLik}
\usage{
ClaytonLogLik(theta, U)
}
\arguments{
\item{theta}{a vector of copula parameter values each in the interval \eqn{[0, \infty)}.
If theta is of length \eqn{< n} then the values in theta will be recycled so
that the parameter vector has length \eqn{n}.}

\item{U}{an \eqn{n x d} matrix of points in \eqn{[0, 1]^d} at which to evaluate the
log-likelihood of the Clayton copula.}
}
\value{
The output is an \eqn{n x 1} vector of values of the log-likelihood. Any
value of the log-likelihood which evaluates to infinity is replaced with -1e9.
}
\description{
A function to compute the Clayton copula log-likelihood.
}
\details{
This function computes the Clayton copula log-likelihood at \eqn{n} points,
where the \eqn{i}th point in \eqn{d} dimensions is specified by the \eqn{i}th
row of the matrix \code{U}. The copula dependence parameter used to compute
the log-likelihood at each point is the \eqn{i}th value in the vector \code{theta}.
If \code{theta} is of length \eqn{< n}, values from \code{theta} are recycled
to produce a vector of length \eqn{n}
}
\examples{
set.seed(100)
n <- 10
d <- 3
theta <- runif(n, min = 0, max = 10)
U <- matrix(runif(n * d), ncol = d)
loglikClayton <- ClaytonLogLik(theta, U)
}
\references{
\insertref{nelsen2006}{MSCopula}
}
