% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claytonCopulaDF.R
\name{dClaytonCopula}
\alias{dClaytonCopula}
\title{dClaytonCopula}
\usage{
dClaytonCopula(U, theta)
}
\arguments{
\item{U}{an \eqn{n x d} matrix giving n points in \eqn{[0, 1]^d} at which to evaluate
the pdf of the Clayton copula}

\item{theta}{a vector of values to be used as the copula dependence
parameter when computing the pdf. If theta is of length \eqn{< n} then the values
in the vector are repeated until the parameter vector is of length \eqn{n}}
}
\value{
The output is an \eqn{n x 1} vector of pdf values. If any point returns
a value that evaluates to \code{Inf} then the function returns the value -1e9.
}
\description{
A function to compute the Clayton copula density.
}
\details{
This function computes the Clayton copula density function at \eqn{n} points,
where the \eqn{i}th point in \eqn{d} dimensions is specified by the \eqn{i}th
row of the matrix \code{U}. The copula dependence parameter used to compute
the density at each point is the \eqn{i}th value in the vector \code{theta}.
If \code{theta} is of length \eqn{< n}, values from \code{theta} are recycled
to produce a vector of length \eqn{n}.
}
\examples{
set.seed(100)
n <- 10
d <- 3
theta <- runif(n, min = 0, max = 10)
U <- matrix(runif(n * d), ncol = d)
pdfClayton <- dClaytonCopula(U, theta)
}
\references{
\insertref{nelsen2006}{MSCopula}
}
