% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankCopulaDF.R
\name{dFrankCopula}
\alias{dFrankCopula}
\title{dFrankCopula}
\usage{
dFrankCopula(U, theta)
}
\arguments{
\item{U}{\itemize{
\item An n x d matrix giving n points in \link{0, 1}^d at which to evaluate
the CDF
}}

\item{theta}{\itemize{
\item A vector of parameters to be used in the computation of the
copula density
}}
}
\description{
dFrankCopula
}
