% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SJCCopulaDF.R
\name{dSJCCopula}
\alias{dSJCCopula}
\title{dSJCCopula}
\usage{
dSJCCopula(U, theta)
}
\arguments{
\item{U}{an \eqn{n x d} matrix giving n points in \eqn{[0, 1]^d} at which to evaluate
the density of the Symmetrized Joe-Clayton copula.}

\item{theta}{either a numeric vector of length 2 or an \eqn{n x 2} matrix of
parameter values in \eqn{[0, 1]}. Values in the first column of \code{theta}
determine the lower tail dependence, while those in the second column determine
the upper tail dependence. If \code{theta} is given as a matrix, values in the
\eqn{i}th row of \code{theta} are used to compute the pdf at the point specified
by row \eqn{i} of \code{U}. If \code{theta} is given as a numeric vector of
length 2 then these values are used to parameterise the density at all points.}
}
\value{
The output is an \eqn{n x 1} vector of pdf values. If any point returns
a value of the density that evaluates to \code{Inf} then NaN is returned.
}
\description{
A function to compute the density function of the Symmetrized Joe-Clayton copula.
}
\details{
This function computes the Symmetrized Joe-Clayton copula density function at \eqn{n}
points, where the \eqn{i}th point in \eqn{d} dimensions is specified by the
\eqn{i}th row of the matrix \code{U}. If \code{theta} is given as a numeric
vector then these parameter values are used to compute the density at all points.
If \code{theta} is a matrix, then the parameters in the \eqn{i}th row of
\code{theta} are used to compute the pdf at the point specified in the \eqn{i}th
row of \code{U}.
}
\examples{
set.seed(100)
n <- 10
d <- 3
theta <- matrix(runif(n * d), ncol = 2)
U <- matrix(runif(n * d), ncol = d)
pdfSymmetrizedJoeClayton <- dSJCCopula(U, theta)
}
\references{
\insertRef{nelsen2006}{MSCopula}
\insertRef{patton2006}{MSCopula}
}
