% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tCopulaDF.R
\name{dtCopula}
\alias{dtCopula}
\title{dtCopula}
\usage{
dtCopula(U, rho, df)
}
\arguments{
\item{U}{\itemize{
\item An n x d matrix giving n points in \link{0, 1}^d at which to evaluate
the CDF
}}

\item{rho}{\itemize{
\item Either a d x d symmetrix, positive definite matrix of
correlation coefficients or a vector of length d*(d-1)/2 specifiying the
values of the correlation coefficients
}}
}
\value{
Y
}
\description{
dtCopula
}
