% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCCopulaDF.R
\name{JCLogLik}
\alias{JCLogLik}
\title{JCLogLik}
\usage{
JCLogLik(theta, U)
}
\arguments{
\item{theta}{either a numeric vector of length 2 or an \eqn{n x 2} matrix of
parameter values in \eqn{[0, 1]}. Values in the first column of \code{theta}
determine the lower tail dependence, while those in the second column determine
the upper tail dependence. If \code{theta} is given as a matrix, the \eqn{i}th row
specifies the point at which the log-likelihood is computed using the data in
row \eqn{i} of the matrix \code{U}.}

\item{U}{an \eqn{n\times d}{n x d} matrix giving n points in \eqn{[0, 1]^d} at which to evaluate
the log-likelihood of the Joe-Clayton copula.}
}
\value{
The output is an \eqn{n x 1} vector of log-likelihood values. If any
point returns a value of the log-likelihood that evaluates to \code{Inf} then
-1e9 is returned.
}
\description{
A function to compute the log-likelihood function of the Joe-Clayton copula.
}
\details{
This function computes the Joe-Clayton copula log-likelihood at \eqn{n}
points, where the \eqn{i}th point in \eqn{d} dimensions is specified by the
\eqn{i}th row of the matrix \code{U}. If \code{theta} is given as a numeric
vector then these parameter values are used to compute the density at all points.
If \code{theta} is a matrix, then the parameters in the \eqn{i}th row of
\code{theta} are used to compute the pdf at the point specified in the \eqn{i}th
row of \code{U}.
}
\examples{
set.seed(100)
n <- 10
d <- 3
theta <- cbind(runif(n), runif(n))
U <- matrix(runif(n * d), ncol = d)
LogLikJoeClayton <- JCLogLik(theta, U)
}
\references{
\insertRef{nelsen2006}{MSCopula}
\insertRef{patton2006}{MSCopula}
}
