% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claytonCopulaDF.R
\name{pClaytonCopula}
\alias{pClaytonCopula}
\title{pClaytonCopula}
\usage{
pClaytonCopula(U, theta)
}
\arguments{
\item{U}{an \eqn{n x d} matrix giving n points in \eqn{[0, 1]^d} at which to evaluate
the cdf of the Clayton copula.}

\item{theta}{scalar dependence parameter for the Clayton copula. Restricted
to the interval \eqn{[0, \infty)}.}
}
\value{
The output is an \eqn{n x 1} vector of cdf values.
}
\description{
A function to compute the distribution function of the Clayton copula.
}
\details{
This function computes the Clayton copula distribution function at \eqn{n}
points, where the \eqn{i}th point in \eqn{d} dimensions is specified by the
\eqn{i}th row of the matrix \code{U}. The copula dependence parameter used to
compute the distribution function at each point is the scalar \code{theta}.
}
\examples{
set.seed(100)
n <- 10
d <- 3
theta <- runif(1, min = 0, max = 10)
U <- matrix(runif(n * d), ncol = d)
cdfClayton <- pClaytonCopula(U, theta)
}
\references{
\insertref{nelsen2006}{MSCopula}
}
