% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankCopulaDF.R
\name{pFrankCopula}
\alias{pFrankCopula}
\title{pFrankCopula}
\usage{
pFrankCopula(U, theta)
}
\arguments{
\item{U}{\itemize{
\item An n x d matrix giving n points in \link{0, 1}^d at which to evaluate
the CDF
}}

\item{theta}{\itemize{
\item The parameter of the Frank copula
}}
}
\description{
pFrankCopula
}
