% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianCopulaDF.R
\name{pGaussianCopula}
\alias{pGaussianCopula}
\title{pGaussianCopula}
\usage{
pGaussianCopula(U, rho)
}
\arguments{
\item{U}{an \eqn{n x d} matrix giving \eqn{n} points in \eqn{[0, 1]^d} at
which to evaluate the distribution function.}

\item{rho}{either a \eqn{d x d} symmetric, positive definite matrix of
correlation coefficients or a vector of length \eqn{d*(d-1)/2}. If \code{rho}
is a vector, the correlation matrix is constructed by filling the upper
triangle of the matrix row-by-row with elements of \code{rho}.}
}
\value{
The output is an \eqn{n x 1} vector of cdf values.
}
\description{
A function to compute the distribution function of the Gaussian copula.
}
\details{
This function computes the Gaussian copula distribution function at \eqn{n}
points, where the \eqn{i}th point in \eqn{d} dimensions is specified by the
\eqn{i}th row of the matrix \code{U}. \code{rho} specifies the correlation
matrix used to parameterise the copula distribution function.
}
\examples{
set.seed(100)
n <- 10
d <- 3
rho <- c(0.4, 0.1, 0.7)
U <- matrix(runif(n * d), ncol = d)
cdfGaussian <- pGaussianCopula(U, rho)
}
\references{
\insertref{nelsen2006}{MSCopula}
}
