% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tCopulaDF.R
\name{ptCopula}
\alias{ptCopula}
\title{ptCopula}
\usage{
ptCopula(U, rho, df)
}
\arguments{
\item{U}{\itemize{
\item An n x d matrix giving n points in \link{0, 1}^d at which to evaluate
the CDF
}}

\item{rho}{\itemize{
\item Either a d x d symmetric, positive definite matrix of
correlation coefficients or a vector of length d*(d-1)/2 specifying the
values of the correlation coefficients
}}
}
\value{
Y
}
\description{
A function to compute the CDF of the t copula at n points in d
dimensions.
}
